Create a specific test class for deathrattle orders
